% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.r
\name{download_data}
\alias{download_data}
\title{An example for writing a function in R using Roxygen (e.g. a variant of Doxygen for
R)
Well, this function is for writing stuff, I suppose that I need to write here the
pourpose of the function.}
\usage{
(id_device, start_date, end_date, sample_rate, logs = FALSE, data_type = 'RAW', file_format= NULL, fields = NULL) }
\arguments{
\item{something}{Something to put}
}
\description{
An example for writing a function in R using Roxygen (e.g. a variant of Doxygen for
R)
Well, this function is for writing stuff, I suppose that I need to write here the
pourpose of the function.
}
\examples{
}
\keyword{Tests,}
\keyword{examples,}
\keyword{many}
