% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.r
\name{download_data}
\alias{download_data}
\title{An example for writing a function in R using Roxygen (e.g. a variant of Doxygen for
R)
Well, this function is for writing stuff, I suppose that I need to write here the
pourpose of the function.}
\usage{
download_data(IdDevice, tmin, tmax, frecuency, token)
}
\arguments{
\item{something}{Something to put}
}
\description{
An example for writing a function in R using Roxygen (e.g. a variant of Doxygen for
R)
Well, this function is for writing stuff, I suppose that I need to write here the
pourpose of the function.
}
\examples{
download_data()
}
\keyword{Tests,}
\keyword{examples,}
\keyword{many}
